# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CCDI Federation Service API",
    "description": "REST API for querying CCDI-DCC  graph database",
    "version": "1.0.0",
    "summary": "Childhood Cancer Data Initiative Federation Service",
    "contact": {
      "name": "CCDI Development Team",
      "email": "support@ccdi.org",
      "url": "https://github.com/CBIIT/ccdi-dcc-federation-service"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1",
      "description": "Local development server (default)"
    },
    {
      "url": "https://dcc-dev.ccdi.cancer.gov/api/v1",
      "description": "Dev Test server"
    }
  ],
  "paths": {
...
    "/namespace": {
      "get": {
        "tags": [
          "Namespace"
        ],
        "summary": "Gets the namespaces known by this server.",
        "description": "Gets the namespaces known by this server.",
        "operationId": "namespace_index",
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespacesResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/namespace/{organization}/{namespace}": {
      "get": {
        "tags": [
          "Namespace"
        ],
        "summary": "Gets the namespace matching the provided name (if it exists).",
        "description": "Gets the namespace matching the provided name (if it exists).",
        "operationId": "namespace_show",
        "parameters": [
          {
            "name": "organization",
            "in": "path",
            "description": "The organization of the namespace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "namespace",
            "in": "path",
            "description": "The name of the namespace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NamespaceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Namespace not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                },
                "example": {
                  "errors": [
                    {
                      "kind": "NotFound",
                      "entity": "Namespaces",
                      "message": "Namespaces not found."
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/organization": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Gets the organizations known by this server.",
        "description": "Gets the organizations known by this server.",
        "operationId": "organization_index",
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      }
    },
    "/organization/{name}": {
      "get": {
        "tags": [
          "Organization"
        ],
        "summary": "Gets the organization matching the provided name (if it exists).",
        "description": "Gets the organization matching the provided name (if it exists).",
        "operationId": "organization_show",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "The name of the organization.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Organization not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                },
                "example": {
                  "errors": [
                    {
                      "kind": "NotFound",
                      "entity": "Organization"
                    }
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorsResponse": {
        "type": "object",
        "description": "Error response envelope containing a list of errors.",
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetail"
            }
          }
        },
        "required": [
          "errors"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "description": "Individual error detail. Fields vary by error kind.",
        "properties": {
          "kind": {
            "type": "string",
            "description": "Error kind identifier",
            "enum": [
              "InvalidRoute",
              "InvalidParameters",
              "NotFound",
              "UnshareableData",
              "UnsupportedField"
            ],
            "example": "NotFound"
          },
          "method": {
            "type": "string",
            "description": "HTTP method (for InvalidRoute errors)",
            "example": "GET"
          },
          "route": {
            "type": "string",
            "description": "Invalid route path (for InvalidRoute errors)",
            "example": "/foobar"
          },
          "parameters": {
            "type": "array",
            "description": "List of invalid parameter names (for InvalidParameters errors)",
            "items": {
              "type": "string"
            },
            "example": [
              "id"
            ]
          },
          "message": {
            "type": "string",
            "description": "Error message explaining what went wrong",
            "example": "The parameter was a non-integer value."
          },
          "entity": {
            "type": "string",
            "description": "Entity type or identifier (for NotFound/UnshareableData errors)",
            "example": "Samples"
          },
          "field": {
            "type": "string",
            "description": "Field name (for UnsupportedField errors)",
            "example": "field"
          }
        },
        "oneOf": [
          {
            "type": "object",
            "required": [
              "kind",
              "method",
              "route",
              "message"
            ]
          },
          {
            "type": "object",
            "required": [
              "kind",
              "parameters",
              "message"
            ]
          },
          {
            "type": "object",
            "required": [
              "kind",
              "entity",
              "message"
            ]
          },
          {
            "type": "object",
            "required": [
              "kind",
              "entity",
              "message"
            ]
          },
          {
            "type": "object",
            "required": [
              "kind",
              "field",
              "message"
            ]
          }
        ]
      },
      "NamespaceEntity": {
        "type": "object",
        "description": "Namespace entity with identifier, description, and metadata.",
        "properties": {
          "id": {
            "$ref": "#/components/schemas/NamespaceIdentifier"
          },
          "description": {
            "type": "string",
            "description": "Human-readable description of the namespace",
            "example": "Namespace for CCDI-DCC phs studies"
          },
          "contact_email": {
            "type": "string",
            "format": "email",
            "description": "Contact email for namespace support",
            "example": "support@ccdi.org"
          },
          "metadata": {
            "type": "object",
            "description": "Additional namespace metadata",
            "nullable": true
          }
        },
        "required": [
          "id",
          "description",
          "contact_email"
        ]
      },
      "NamespaceIdentifier": {
        "type": "object",
        "description": "Namespace identifier composed of organization and name.",
        "properties": {
          "organization": {
            "type": "string",
            "description": "Organization identifier",
            "example": "CCDI-DCC",
            "const": "CCDI-DCC"
          },
          "name": {
            "type": "string",
            "description": "Namespace name (study identifier)",
            "example": "phs"
          }
        },
        "required": [
          "organization",
          "name"
        ]
      },
      "OrganizationEntity": {
        "type": "object",
        "description": "Organization entity with identifier, name, and metadata.",
        "properties": {
          "identifier": {
            "type": "string",
            "description": "Organization identifier",
            "example": "CCDI-DCC",
            "const": "CCDI-DCC"
          },
          "name": {
            "type": "string",
            "description": "Human-readable organization name",
            "example": "Childhood Cancer Data Initiative Data Coordinating Center"
          },
          "metadata": {
            "type": "object",
            "description": "Additional organization metadata",
            "nullable": true
          }
        },
        "required": [
          "identifier",
          "name"
        ]
      },
      "NamespacesResponse": {
        "type": "object",
        "description": "Response containing a list of namespaces.",
        "properties": {
          "namespaces": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NamespaceEntity"
            }
          }
        },
        "required": [
          "namespaces"
        ]
      },
      "NamespaceResponse": {
        "type": "object",
        "description": "Response containing a single namespace.",
        "properties": {
          "namespace": {
            "$ref": "#/components/schemas/NamespaceEntity"
          }
        },
        "required": [
          "namespace"
        ]
      },
      "OrganizationsResponse": {
        "type": "object",
        "description": "Response containing a list of organizations.",
        "properties": {
          "organizations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrganizationEntity"
            }
          }
        },
        "required": [
          "organizations"
        ]
      },
      "OrganizationResponse": {
        "type": "object",
        "description": "Response containing a single organization.",
        "properties": {
          "organization": {
            "$ref": "#/components/schemas/OrganizationEntity"
          }
        },
        "required": [
          "organization"
        ]
      },
...
    }
  },
...
}